move hrp to chain_cosmos
