+++ crates/gem_cosmos/src/converter.rs
@@ -12,14 +12,14 @@ pub fn convert_cosmos_address(address: &str, hrp: &str) -> Result<String, anyhow
 #[cfg(test)]
 mod tests {
     use super::*;
-    use primitives::Chain;
+    use primitives::CosmosChain;
 
     #[test]
     fn test_cosmos_osmosis_convert() {
         let address = "cosmos1fxygpgus4nd5jmfl5j7fh5y8hyy53z8u95dzx7";
         let expected = "osmo1fxygpgus4nd5jmfl5j7fh5y8hyy53z8ud07jsv";
 
-        let output_address = convert_cosmos_address(address, Chain::Osmosis.hrp()).unwrap();
+        let output_address = convert_cosmos_address(address, CosmosChain::Osmosis.hrp()).unwrap();
         assert_eq!(expected, output_address);
     }
 
@@ -28,7 +28,7 @@ mod tests {
         let address = "inj1kgq0kzzatjh0lzv73n0nyvyen6npladdz62dtr";
         let expected = "dym1kgq0kzzatjh0lzv73n0nyvyen6npladd6w30u4";
 
-        let output_address = convert_cosmos_address(address, Chain::Dymension.hrp()).unwrap();
+        let output_address = convert_cosmos_address(address, CosmosChain::Dymension.hrp()).unwrap();
         assert_eq!(expected, output_address);
     }
 
@@ -37,7 +37,7 @@ mod tests {
         let address = "cosmos1fxygpgus4nd5jmfl5j7fh5y8hyy53z8u95dzx7";
         let expected = "saga1fxygpgus4nd5jmfl5j7fh5y8hyy53z8um85spc";
 
-        let output_address = convert_cosmos_address(address, Chain::Saga.hrp()).unwrap();
+        let output_address = convert_cosmos_address(address, CosmosChain::Saga.hrp()).unwrap();
         assert_eq!(expected, output_address);
     }
