+++ crates/primitives/src/chain_cosmos.rs
@@ -1,6 +1,12 @@
-#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
+use serde::{Deserialize, Serialize};
+use strum::EnumIter;
+use strum_macros::{AsRefStr, EnumString};
+use typeshare::typeshare;
+
+#[derive(Copy, Clone, Debug, Serialize, Deserialize, EnumIter, AsRefStr, EnumString)]
 #[typeshare(swift = "Equatable, Codable, CaseIterable")]
 #[serde(rename_all = "lowercase")]
+#[strum(serialize_all = "lowercase")]
 pub enum CosmosChain {
     Cosmos,
     Osmosis,
@@ -13,6 +19,7 @@ pub enum CosmosChain {
     Saga,
 }
 
+#[allow(non_camel_case_types)]
 #[typeshare(swift = "Equatable, Codable, CaseIterable")]
 pub enum CosmosDenom {
     rune,
@@ -24,4 +31,20 @@ pub enum CosmosDenom {
     uusdc,
     adym,
     usaga,
+}
+
+impl CosmosChain {
+    pub fn hrp(&self) -> &str {
+        match self {
+            Self::Cosmos => "cosmos",
+            Self::Osmosis => "osmo",
+            Self::Celestia => "celestia",
+            Self::Noble => "noble",
+            Self::Injective => "inj",
+            Self::Sei => "sei",
+            Self::Thorchain => "thor",
+            Self::Dymension => "dym",
+            Self::Saga => "saga",
+        }
+    }
 }
