+++ crates/primitives/src/lib.rs
@@ -2,6 +2,8 @@
 
 pub mod chain;
 pub use self::chain::Chain;
+pub mod chain_cosmos;
+pub use self::chain_cosmos::CosmosChain;
 pub mod chain_type;
 pub use self::chain_type::ChainType;
 pub mod chain_evm;
